@@ -27,7 +27,6 @@ classifiers = [
   "Typing :: Typed",
 ]
 dependencies = [
-  "msgspec~=0.19",
 ]
 [optional-dependencies]
 requests = ["requests~=2.32"]
@@ -91,4 +90,4 @@ exclude_lines = [
 default-args = ["tests/__init__.py"]
 
 [[tool.hatch.envs.hatch-test.matrix]]
-python = ["3.13", "3.12"]
+python = ["3.14", "3.13", "3.12"]
